requirements.txt
annotated-types==0.7.0
anyio==4.9.0
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.12
idna==3.10
pyasn1==0.4.8
pydantic==2.11.2
pydantic-core==2.33.1
python-jose==3.4.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing-extensions==4.13.1
typing-inspection==0.4.0